Moved generic components to separate files
