+++ src/control-flow/cfg-python.ts
@@ -1,190 +1,13 @@
-import { MultiDirectedGraph } from "graphology";
 import Parser from "web-tree-sitter";
 import {
-  BlockHandler,
   type BasicBlock,
+  BlockHandler,
   type BuilderOptions,
   type CFG,
-  type CFGGraph,
-  type Cluster,
-  type ClusterId,
-  type ClusterType,
-  type EdgeType,
-  type NodeType,
 } from "./cfg-defs";
-
-class Builder {
-  private graph: CFGGraph = new MultiDirectedGraph();
-  private nodeId: number = 0;
-  private clusterId: ClusterId = 0;
-  private activeClusters: Cluster[] = [];
-
-  private startCluster(type: ClusterType): Cluster {
-    const parent =
-      this.activeClusters.length === 0
-        ? undefined
-        : this.activeClusters[this.activeClusters.length - 1];
-    const cluster = {
-      id: this.clusterId++,
-      type,
-      parent,
-      depth: this.activeClusters.length + 1,
-    };
-    this.activeClusters.push(cluster);
-    return cluster;
-  }
-  private endCluster(_cluster: Cluster) {
-    // We assume that all clusters form a stack.
-    this.activeClusters.pop();
-  }
-
-  public withCluster<T>(type: ClusterType, fn: (cluster: Cluster) => T): T {
-    const cluster = this.startCluster(type);
-    try {
-      return fn(cluster);
-    } finally {
-      this.endCluster(cluster);
-    }
-  }
-  public addNode(type: NodeType, code: string, lines: number = 1): string {
-    const id = `node${this.nodeId++}`;
-    const cluster = this.activeClusters[this.activeClusters.length - 1];
-    this.graph.addNode(id, {
-      type,
-      code,
-      lines,
-      markers: [],
-      cluster,
-    });
-    return id;
-  }
-
-  public cloneNode(node: string, overrides?: { cluster: Cluster }): string {
-    const id = `node${this.nodeId++}`;
-    const originalAttrs = this.graph.getNodeAttributes(node);
-    const nodeAttrs = structuredClone(originalAttrs);
-    nodeAttrs.cluster = originalAttrs.cluster;
-    Object.assign(nodeAttrs, overrides);
-    this.graph.addNode(id, nodeAttrs);
-    return id;
-  }
-
-  public addMarker(node: string, marker: string) {
-    this.graph.getNodeAttributes(node).markers.push(marker);
-  }
-
-  public addEdge(
-    source: string,
-    target: string,
-    type: EdgeType = "regular",
-  ): void {
-    if (!this.graph.hasEdge(source, target)) {
-      this.graph.addEdge(source, target, { type });
-    }
-  }
-
-  public getGraph(): CFGGraph {
-    return this.graph;
-  }
-}
-
-function matchQuery(
-  syntax: Parser.SyntaxNode,
-  mainName: string,
-  queryString: string,
-) {
-  const language = syntax.tree.getLanguage();
-  const query = language.query(queryString);
-  const matches = query.matches(syntax);
-  const match = (() => {
-    for (const match of matches) {
-      for (const capture of match.captures) {
-        if (capture.name === mainName && capture.node.id === syntax.id) {
-          return match;
-        }
-      }
-    }
-    throw new Error("No match found!");
-  })();
-  return match;
-}
-
-function matchExistsIn(
-  syntax: Parser.SyntaxNode,
-  mainName: string,
-  queryString: string,
-): boolean {
-  const language = syntax.tree.getLanguage();
-  const query = language.query(queryString);
-  const matches = query.matches(syntax);
-  return matches.length > 0;
-}
-
-function getSyntax(
-  match: Parser.QueryMatch,
-  name: string,
-): Parser.SyntaxNode | undefined {
-  return getSyntaxMany(match, name)[0];
-}
-
-function getSyntaxMany(
-  match: Parser.QueryMatch,
-  name: string,
-): Parser.SyntaxNode[] {
-  return match.captures
-    .filter((capture) => capture.name === name)
-    .map((capture) => capture.node);
-}
-
-class BlockMatcher {
-  private blockHandler: BlockHandler = new BlockHandler();
-  private processBlock: (syntax: Parser.SyntaxNode | null) => BasicBlock;
-  public update = this.blockHandler.update.bind(this.blockHandler);
-  private match: Parser.QueryMatch;
-
-  constructor(
-    processBlock: BlockMatcher["processBlock"],
-    syntax: Parser.SyntaxNode,
-    mainName: string,
-    query: string,
-  ) {
-    this.processBlock = processBlock;
-    this.match = matchQuery(syntax, mainName, query);
-  }
-
-  public getSyntax(name: string): ReturnType<typeof getSyntax> {
-    return getSyntax(this.match, name);
-  }
-
-  public getSyntaxMany(name: string): ReturnType<typeof getSyntaxMany> {
-    return getSyntaxMany(this.match, name);
-  }
-
-  public getBlock(syntax: Parser.SyntaxNode | null | undefined) {
-    return syntax ? this.blockHandler.update(this.processBlock(syntax)) : null;
-  }
-
-  public get state() {
-    return this.blockHandler;
-  }
-}
-
-type StatementHandler = (
-  syntax: Parser.SyntaxNode,
-  builder: Builder,
-  match: (
-    syntax: Parser.SyntaxNode,
-    mainName: string,
-    query: string,
-  ) => BlockMatcher,
-  options: BuilderOptions,
-) => BasicBlock;
-
-type StatementHandlers = {
-  block: string;
-  named: { [key: string]: StatementHandler };
-  default: StatementHandler;
-};
+import { Builder } from "./builder.ts";
+import { BlockMatcher, matchExistsIn } from "./block-matcher.ts";
+import type { StatementHandlers } from "./statement-handlers.ts";
 
 const pythonStatementHandlers: StatementHandlers = {
   block: "block",
